% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjMatrix.R
\name{adjMatrix}
\alias{adjMatrix}
\title{Generates adjacency matrix}
\usage{
adjMatrix(vertices, outfile = tempfile(fileext = .txt))
}
\arguments{
\item{vertices}{matrix of vertices (n x d)}

\item{outfile}{output file}
}
\description{
Generates adjacency matrix
}
\examples{
v <- rbind(
 c( -5, -5,  16 )
, c( -5,  8,   3 )
, c(  4, -1,   3 )
, c(  4, -5,   7 )
, c(  4, -1, -10 )
, c(  4, -5, -10 )
, c( -5,  8, -10 )
, c( -5, -5, -10 ))
adjMatrix(v)
}
